+++ src/batchStub.ts
@@ -19,6 +19,9 @@ class PushStub implements BatchSDK.PushModule {
   }
 
   public registerForRemoteNotifications() {}
+  public refreshToken() {}
+  public requestNotificationAuthorization() {}
+  public requestProvisionalNotificationAuthorization() {}
   public setiOSShowForegroundNotifications(_showForeground: boolean) {}
   public setAndroidNotificationTypes(_notifTypes: unknown) {}
   public setiOSNotificationTypes(_notifTypes: unknown) {}